+++ cli/tests/test_immutable_commits.rs
@@ -380,7 +380,7 @@ fn test_rewrite_immutable_commands() {
     Hint: For more information, see:
           - https://jj-vcs.github.io/jj/latest/config/#set-of-immutable-commits
           - `jj help -k config`, "Set of immutable commits"
-    Hint: This operation would rewrite 2 immutable commits.
+    Hint: This operation would rewrite 1 immutable commits.
     [EOF]
     [exit status: 1]
     "#);